v2.0.1
Update install script.

After 'fdroidserver' is installed into a 'virtualenv',
perform a series of cleanup tasks,
which remove absolute filepaths from various scripts and executables
within the directory:
  ./virtualenv/fdroid-server/Scripts

Previously, the 'virtualenv' directory included in releases
would only run if extracted to:
  E:/fd

Now, the 'virtualenv' directory is fully portable.
The 'bin' and 'scripts' in this repo can be saved to any directory path,
for example:
  C:/fdroidserver-for-windows
and the 'virtualenv' directory included in releases
can be extracted to this same directory.. and it will work.